Fixes silent failures for a few exporters
One example is if a user tries to export a file to an invalid file or
is not permitted to write to it.  For the FFmpeg, MP2, and OGG exporters
this error would not be flagged and to the user, it would look like the
file was exported.

These now throw up an error dialog to left the user know.